refactor: move `GetWorkspace.Resource` to `RAWebServer.Resource`
